+++ index.html
@@ -226,4 +226,4 @@
 </body>
 </html>
 
-<script src='main.js' type='module'></script>
+<script src='./js/main.js' type='module'></script>
